requirements.txt
astroid==2.5.1
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==1.1.2
Flask-WTF==0.15.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
python-dotenv==0.19.1
Werkzeug==2.0.2
wrapt==1.12.1
WTForms==2.3.3
